Add the motivation for having good commit messages
A commit message serves to some purposes, and having a shared
understanding of those purposes is the first step to have good commit
messages.

The text that this commit will add tries to explain which purposes a
commit message serves and based on them how a commit message should be.

Co-authored-by: Luís Cobucci <luis.cobucci@castoredc.com>
Signed-off-by: Henrique Moody <henrique.moody@castoredc.com>