Add the composition of a commit message
Understanding how a commit message is composed is one of the most basic
things to know about commit messages. Anyone can find that information
in the manpage of the "git commit" [1].

This commit will add such information, but at the moment there is an
example of a commit message that is not composed by Subject and Body.
Such a commit message should appear in the upcoming commits.

[1]: https://www.kernel.org/pub/software/scm/git/docs/git-commit.html#_discussion

Co-authored-by: Luís Cobucci <luis.cobucci@castoredc.com>
Signed-off-by: Henrique Moody <henrique.moody@castoredc.com>