add the third guideline for a good commit message
A commit message is a document. At school, they teach us at a young age
that we should capitalize the first letter of a sentence and it should
not be different in a commit message.

Besides, if you compare this commit message with the other ones, it will
look inconsistent since we capitalized all the other ones.

Signed-off-by: Henrique Moody <henrique.moody@castoredc.com>