Add the #5 guideline for a good commit message
Besides the fact that we have 50 characters for the Subject line [1] and
that we want to avoid noise on that line [2], having the issue number in
the commit message does not serve the purpose of a repository.

An issue tracker, like Jira, has the purpose of cataloging requirement
of a feature, a big-fix, etc. The issue does not explain why the
developer changed or added the code in a specific way.

Another factor is that a repository should be self-contained, especially
if we are talking about a distributed version control system like Git.
Having to consult a different source is, at least, a pity.

If someone wants to add the issue as a piece of extra information,
adding a link at the bottom of the body that will be more suitable and
easy, so other people developers can click on it and go to the issue
tracker page.

[1]: e42560fa8ca0392dcb9e233122c79d41e4e31e85
[2]: eab01507e95b9b90052272e7d47181995b4b5e39

Co-authored-by: Luís Cobucci <luis.cobucci@castoredc.com>
Signed-off-by: Henrique Moody <henrique.moody@castoredc.com>