Add the eighth guideline for a good commit message
Developers understand code, and it is often easy to understand WHAT a
commit adds or changes. The information that the commit does not give us
out of the box is WHY.

Sometimes a piece of code does not make sense and when someone changes
it the application breaks. Then someone finds out who is the author of
that piece of code, but this person is on vacation or even left the
team or company. Developers can reduce situations like that by having
good commit messages.

That also helps with code reviews to understand the reasoning of the
author of a commit.

Co-authored-by: Luís Cobucci <luis.cobucci@castoredc.com>
Signed-off-by: Henrique Moody <henrique.moody@castoredc.com>